# microservices/product-management/src/main/java/uk/ac/york/eng2/products/dto/ProductCreateDTO.java
package uk.ac.york.eng2.products.dto;

import io.micronaut.serde.annotation.Serdeable;

import java.math.BigDecimal;
import java.util.HashSet;
import java.util.Set;

@Serdeable
public class ProductCreateDTO {
    private String name;
    private BigDecimal unitPrice;
    private Set<String> tags = new HashSet<>();

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public BigDecimal getUnitPrice() {
        return unitPrice;
    }

    public void setUnitPrice(BigDecimal unitPrice) {
        this.unitPrice = unitPrice;
    }

    public Set<String> getTags() {
        return tags;
    }

    public void setTags(Set<String> tags) {
        this.tags = tags;
    }
}